180 introduce ClassDefinitionFactory
