+++ src/Kontent.Ai.ModelGenerator.Core/Contract/IClassDefinitionFactory.cs
@@ -1,0 +1,8 @@
+﻿using Kontent.Ai.ModelGenerator.Core.Common;
+
+namespace Kontent.Ai.ModelGenerator.Core.Contract;
+
+public interface IClassDefinitionFactory
+{
+    ClassDefinition CreateClassDefinition(string codename);
+}
